moving the app navigation bar to the design system
