+++ lib/presentation/design_system/widget/app_navigation_bar.dart
@@ -6,7 +6,7 @@ import 'package:share_space/presentation/screen/home/home_screen.dart';
 import 'package:share_space/presentation/screen/my_account/my_account_screen.dart';
 import 'package:share_space/presentation/screen/search/search_screen.dart';
 
-import '../design_system/theme/app_theme.dart';
+import '../theme/app_theme.dart';
 
 class AppNavigationBar extends StatefulWidget {
   const AppNavigationBar({super.key});
