+++ lib/presentation/routes/routes.dart
@@ -8,7 +8,7 @@ import 'package:share_space/presentation/screen/my_account/my_account_screen.dar
 import 'package:share_space/presentation/screen/onboarding/onboarding_screen.dart';
 import 'package:share_space/presentation/screen/room_details/room_details_screen.dart';
 import 'package:share_space/presentation/screen/search/search_screen.dart';
-import 'package:share_space/presentation/widgets/app_navigation_bar.dart';
+import 'package:share_space/presentation/design_system/widget/app_navigation_bar.dart';
 import '../screen/create_account/create_account_screen.dart';
 
 class Routes {
